Boss bullet adjustment

+++ Boss.h
@@ -132,7 +132,7 @@ public:
     static float bulletSpeed_;
     static Vector3 knockbackdis;
     static int damage_;
-    static int bombBaseDamage_;
+    static float bombBaseDamage_;
     static int shotCount;
 
 };
